Created order-service and configured it to eureka server.

+++ services/order/src/main/java/com/yazdahni/order/orderline/OrderLine.java
@@ -22,6 +22,6 @@ public class OrderLine {
 
     private Integer productId;
 
-    private Integer quantity;
+    private double quantity;
 
 }
